requirements.txt
flask~=2.2.3
python-decouple==3.8
